Change to Arcadam syntax
In general, I believe Arcadam succeeds in being more clear and regular than Markdown.

Some of the wins (all attributable to Asciidoc):
- Block title as standard
- Block attributes as standard
- Link to a heading
- Easier to escape `
- Add term to index without repeating it
- Code callouts (possibly contentious)
- Description list

Arcadam requires all links to be wrapped in brackets. I think it actually helps readability. It doesn't support inline footnotes, but with automatic numbering that is not a big issue and could help readability. The only list item number it allows is 1, but I noticed this document follows the exact same convention.

Some notes:
* Normally a link to a heading is written [Heading text]() and the processor looks for headings in the current document. But in this case there are multiple documents, so it requires an extension to handle that. Similarly the bibliography is in a different file.

* The output actually uses numbers for internal links. By default [Heading text]() should display as "Heading text" and [](#anchor-name) should display as "#anchor-name". But with blogdown it appears as something like "1.8". That would be another extension.

* I followed the existing convention of starting the anchor name with a type of target and ":", for example in "#fig:figure-name" or in "#tab:table-name". I am thinking of stealing the idea.