Change of plans for indented text.
I tried it on real-world examples and did not see the expected benefits of alternating normal paragraph and code. Let us scrap the idea.

An indented block is code if it's not a list item and comes after non-indented text, that's it. As simple as that.

Also, I found how to indent things without Markdown making wild guesses - just indent two spaces at a time. Or if what I really want is a code block, four spaces.

That makes everything more readable without losing support from Markdown editors.