Simple table syntax; {#target} syntax; fix image button
The question mark in the URL wrapper is now required: [?](address)

No more talk about having an empty image caption used to add a link to the image.

The submit button has a default label anyway, so it didn't really make sense.

Form options are a bit of an issue for translators since the value gets sent to the target. That makes a bullet list mostly unsuitable.

Putting the field id in the parameters avoids writing one more line to set it.