Block visibility toggle
Syntax:

[@show]: +
[@hide]: -
[@expand]: >
[@collapse]: ^

Note that the custom marker is considered as a link reference in Markdown with this syntax and it does not appear in the output.

The idea is that e.g. a footnote is hidden because the contents are shown at a different place than where they appear in the document (at the bottom of the page).