Refinement of text formatting rules
The Asciidoctor rules don't look much saner than Markdown so I am making my own rules.

One thing I dislike with Asciidoc is the `+mono+` format and hopefully that avoids it.

The difficult part will be to reconcile:
* Format markers cannot cross a
  boundary.
* The formatter ignores a single format marker enclosed by the same marker which is doubled

I don't believe it is ambiguous but could be a headache to implement efficiently.